feat(toolchain): 集成 ty 类型检查并修复全部 51 处类型问题
选型: ty (Astral/Rust) 而非 pyright —— 与 uv+ruff 同工具链、无需 Node、更快、CI 更简。
pyproject 增加 ty dev 依赖与 [tool.ty] 配置；CI 增加 ty check 步骤。

类型问题修复（ty 51 → 0）:
- 真实缺陷: SerialConfig.stop_bits 改 float(支持 1.5 停止位); set_protocol 接受 None;
  multi_port_dialog 的 config/get_config 改 Optional; wait_response/template_dir/
  LogConfig/ThemeMode 等 Optional 默认值补全
- 健壮性: 表格 item()/按钮/菜单/状态栏/表头等 PyQt6 可选返回值增加 None 守卫;
  closeEvent 参数名匹配 Qt 桩(a0)并守卫 accept; script_engine 上下文守卫
- 设计: 6 个协议 build_frame 特化签名以 # type: ignore 显式标注(基类 **kwargs 契约的
  刻意特化), 保留各协议类型化的构建 API

验证: ty/ruff All checks passed, 21 测试通过

+++ src/core/serial_manager.py
@@ -25,7 +25,7 @@ class SerialConfig:
     port: str
     baudrate: int = 115200
     data_bits: int = 8
-    stop_bits: int = 1
+    stop_bits: float = 1
     parity: str = 'N'  # N, E, O, M, S
     flow_control: str = 'None'  # None, RTS/CTS, XON/XOFF
     timeout: float = 0.1
